requirements.txt
# install pytorch 2.5, torchvision, torchao, grpc

torch==2.8.0
torchvision
torchao
grpcio
grpcio-tools
numpy
pandas
scikit-learn
ray

